+++ TensorWeaver/OutputData/Classification.cs
@@ -2,10 +2,10 @@ namespace TensorWeaver.OutputData;
 
 public readonly struct Classification
 {
-	public byte ClassId { get; }
+	public ushort ClassId { get; }
 	public float Confidence { get; }
 
-	public Classification(byte classId, float confidence)
+	public Classification(ushort classId, float confidence)
 	{
 		ClassId = classId;
 		Confidence = confidence;
