+++ TensorWeaver/OutputData/Detection.cs
@@ -4,7 +4,7 @@ public readonly struct Detection
 {
 	public Classification Classification { get; }
 	public Bounding Bounding { get; }
-	public byte ClassId => Classification.ClassId;
+	public ushort ClassId => Classification.ClassId;
 	public float Confidence => Classification.Confidence;
 
 	public ushort Index { get; }
